// src/infrastructure/settings/dto/user.profile.dto.ts
export type FarmData = {
  farmName: string;
  farmType: string;
  farmDescription: string;
  farmSize: number;

  country: string;
  region: string;

  firstName: string;
  lastName: string;
  emailAddress: string;
  phoneNumber: string;
  bioInformation: string;

  county: string;
  subCounty: string;
  ward: string;
  farmLocation: string;
  farmCoordinates: string;
};